pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Urkel",email = "abmichael109@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "dj-database-url (>=3.0.0,<4.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "lxml (>=5.4.0,<6.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "daphne (>=4.2.0,<5.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "removebg (>=0.4,<0.5)",
    "opencv-python-headless (>=4.12.0.88,<5.0.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "playwright (>=1.56.0,<2.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "django-storages[s3] (>=1.14.6,<2.0.0)",
    "b2sdk (>=2.10.2,<3.0.0)",
    "boto3 (>=1.42.44,<2.0.0)",
    "rembg (>=2.0.72,<3.0.0)",
    "onnxruntime (>=1.24.3,<2.0.0)",
    "openai (>=2.30.0,<3.0.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
